Fixed shaders to draw uniform colors
[Description]
or MSAA, the subsample of the pixel may be in the triangle,
but the center of the pixel is outside the triangle.
This pixel depth may be out of the depth range
and gl_FragCoord.z is negative, especially when the
triangle is narrow and the depth gradient is large.
This will draw an unexpected black for the pixel
after the calculation.

Affects: dEQP-GLES3.functional.occlusion_query.*

Components: GLES

Bug: 259510331
Test: deqp pass

Change-Id: I14894a03cf02542f133760f38675d4f52d25aee1