Mark tests as "Assumption Failure" that are not required by the deqp level
Context: These tests were previously silently marked as succeed which
makes it harder to test new tests on old phones e.g. disabled tests that
reenabled  in a later year.

Also, Update Unit tests to expect the new return values. Original tests
written in a way to expect success when a test should not run(ag/543890)
which is no more the case.

Test: dEQP-VK.clipping.clip_volume and CTSDeqpRunnerTests
Bug: 309636762
Change-Id: I1074593e1f6b1cbbc88cc54cb6f8f62f87cc4a84