Add mesh and task shader to s_resultItemMap for result parsing
Tests with mesh and task shader can fail when parsing the QPA log file
even the actual tests itself are passing. The failed message reads as
following:
  "InternalError: Could not map 'MeshShader' to result item type".

This change add the MeshShader and TaskShader items to s_resultItemMap
so that tests would not fail at log parsing.

This is a cherrypick from Khronos CL 12474
Done manually because formating was causiong merge conflics.
Test: atest CtsDeqpTestCases
Bug: 293388966

Change-Id: I604e556e3882ff390445bee5509de8245e9ded74